backend/prj/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
billiard==3.6.4.0
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-appconf==1.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
django-webpack-loader==0.7.0
djangochannelsrestframework==0.2.2
djangorestframework==3.12.4
docopt==0.6.2
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
google-api-core==1.26.3
google-api-python-client==2.3.0
google-api-python-client-helpers==1.2.6
google-apis-oauth-django==0.1.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
implicit==0.4.4
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
names-dataset==2.0.1
nest-asyncio==1.5.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pickleshare==0.7.5
Pillow==8.1.2
pip==21.1.1
pipreqs==0.4.10
pluggy==0.13.1
positional==1.2.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-surprise==1.1.1
scipy==1.6.3
Send2Trash==1.5.0
service-identity==18.1.0
setuptools==56.0.0
six==1.15.0
sqlparse==0.4.1
tenacity==7.0.0
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
tornado==6.1
tqdm==4.60.0
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
wrapt==1.12.1
yarg==0.1.9
zope.interface==5.2.0
